+++ CHANGELOG.md
@@ -13,9 +13,11 @@
 
 ## Master
 
+- Fix Netlify integration when repo url includes an access token - [@imorente][]
+
 # 6.1.0
 
-- Add CI integration for Netlify - imorente
+- Add CI integration for Netlify - [@imorente][]
 
 # 6.0.7
 
@@ -1382,3 +1384,4 @@ Not usable for others, only stubs of classes etc. - [@orta][]
 [@johansteffner]: https://github.com/johansteffner
 [@sgtcoolguy]: https://github.com/sgtcoolguy
 [@patrickkempff]: https://github.com/patrickkempff
+[@imorente]: https://github.com/imorente