+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 
 ## Master
 
+- Improves debug logging, messaging on CI about status updates - [@orta][]
+- Better detection of json results: {} from a sub-process - [@orta][]
+- CLI Args are actually passed to subprocess instead of an empty object - [@orta][]
 - Fix Netlify integration when repo url includes an access token - [@imorente][]
 
 # 6.1.0