+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ## Master
 
+# 6.1.1
+
 - Improves debug logging, messaging on CI about status updates - [@orta][]
 - Better detection of json results: {} from a sub-process - [@orta][]
 - CLI Args are actually passed to subprocess instead of an empty object - [@orta][]